Moved the list of hotline to HotlineDropdownPicker component
